packages/python/graphql-server/requirements.txt
black
fastapi
pre-commit
strawberry-graphql[fastapi]
uvicorn